Modified the simple test to use integer values. It works...hooray!

+++ Our Project/corewarV3/corewarV3_opencl.cpp
@@ -133,17 +133,17 @@ cl_program CreateProgram(cl_context context, cl_device_id device,
 }
 
 bool CreateMemObjects(cl_context context, cl_mem memObjects[3],
-                      float *a, float *b)
+                      int *a, int *b)
 {
     memObjects[0] = clCreateBuffer(context,
                                    CL_MEM_READ_ONLY | CL_MEM_COPY_HOST_PTR,
-                                   sizeof(float) * ARRAY_SIZE, a, NULL);
+                                   sizeof(int) * ARRAY_SIZE, a, NULL);
     memObjects[1] = clCreateBuffer(context,
                                    CL_MEM_READ_ONLY | CL_MEM_COPY_HOST_PTR,
-                                   sizeof(float) * ARRAY_SIZE, b, NULL);
+                                   sizeof(int) * ARRAY_SIZE, b, NULL);
     memObjects[2] = clCreateBuffer(context,
                                    CL_MEM_READ_WRITE,
-                                   sizeof(float) * ARRAY_SIZE, NULL, NULL);
+                                   sizeof(int) * ARRAY_SIZE, NULL, NULL);
     if(memObjects[0] == NULL || memObjects[1] == NULL || memObjects[2] == NULL)
     {
         printf("Error creating memory objects. %f %f %f\n", memObjects[0],
@@ -208,9 +208,9 @@ int main(int argc, char** argv)
         printf("Failed to create kernel\n");
     }
 
-    float result[ARRAY_SIZE];
-    float a[ARRAY_SIZE];
-    float b[ARRAY_SIZE];
+    int result[ARRAY_SIZE];
+    int a[ARRAY_SIZE];
+    int b[ARRAY_SIZE];
     for(int i = 0; i < ARRAY_SIZE; i++)
     {
         a[i] = i;
@@ -248,7 +248,7 @@ int main(int argc, char** argv)
     }
 
     errNum = clEnqueueReadBuffer(commandQueue, memObjects[2], CL_TRUE, 0,
-            ARRAY_SIZE * sizeof(float), result, 0, NULL, NULL);
+            ARRAY_SIZE * sizeof(int), result, 0, NULL, NULL);
     if(errNum != CL_SUCCESS)
     {
         printf("Error reading result buffer.\n");
